Align microbiome reads to a set of reference libraries

This is the main MetaScope target library mapping function, using Rsubread and multiple libraries. Aligns to each library separately, filters unmapped reads from each file, and then merges and sorts the .bam files from each library into one output file. If desired, output can be passed to `filter_host()` to remove reads that also map to filter library genomes.

Usage

align_target(
  reads,
  libs,
  lib_dir = NULL,
  project_name = tools::file_path_sans_ext(reads),
  settings = align_details
)

Arguments

reads

Location of the .fastq file to align

libs

A vector of character strings giving the basenames of the Subread index files for alignment. These should be located in the same directory and the current directory if lib_dir=NULL.

lib_dir

path to the library index files (all libraries should be here)

project_name

A name for the project, which names the output .bam file (e.g. project_name.bam). Defaults to the basename of the reads file.

settings

A named list specifying alignment parameters for the Rsubread::align() function, which is called inside align_target(). Elements should include type, nthreads, maxMismatches, nsubreads, phredOffset, unique, and nBestLocations. Descriptions of these parameters are available under ?Rsubread::align. Defaults to the global align_details object.

Value

This function writes a merged and sorted .bam file after aligning to all reference libraries given, along with a summary report file, to the user's working directory. The function also outputs the new .bam filename.

Examples

#### Align example reads to an example reference library using Rsubread

## Create object with path to example reference library
refPath <- system.file("extdata","target.fasta", package = "MetaScope")

## Copy the example reference library to the current directory
file.copy(from = refPath, to = file.path(".", "target.fasta"))
#> [1] FALSE

## Make and align to a single reference library
mk_subread_index("target.fasta")
#> [1] "Library size is 0 GBs. Building the Rsubread index"
#> 
#>         ==========     _____ _    _ ____  _____  ______          _____  
#>         =====         / ____| |  | |  _ \|  __ \|  ____|   /\   |  __ \ 
#>           =====      | (___ | |  | | |_) | |__) | |__     /  \  | |  | |
#>             ====      \___ \| |  | |  _ <|  _  /|  __|   / /\ \ | |  | |
#>               ====    ____) | |__| | |_) | | \ \| |____ / ____ \| |__| |
#>         ==========   |_____/ \____/|____/|_|  \_\______/_/    \_\_____/
#>        Rsubread 2.8.1
#> 
#> //================================= setting ==================================\\
#> ||                                                                            ||
#> ||                Index name : target                                         ||
#> ||               Index space : base space                                     ||
#> ||               Index split : no-split                                       ||
#> ||          Repeat threshold : 100 repeats                                    ||
#> ||              Gapped index : no                                             ||
#> ||                                                                            ||
#> ||       Free / total memory : 318.3GB / 376.4GB                              ||
#> ||                                                                            ||
#> ||               Input files : 1 file in total                                ||
#> ||                             o target.fasta                                 ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> //================================= Running ==================================\\
#> ||                                                                            ||
#> || Check the integrity of provided reference sequences ...                    ||
#> || No format issues were found                                                ||
#> || Scan uninformative subreads in reference sequences ...                     ||
#> || 45 uninformative subreads were found.                                      ||
#> || These subreads were excluded from index building.                          ||
#> || Estimate the index size...                                                 ||
#> ||    8%,   0 mins elapsed, rate=4059.9k bps/s                                ||
#> ||   16%,   0 mins elapsed, rate=4996.7k bps/s                                ||
#> ||   24%,   0 mins elapsed, rate=5400.2k bps/s                                ||
#> ||   33%,   0 mins elapsed, rate=5220.0k bps/s                                ||
#> ||   41%,   0 mins elapsed, rate=5459.0k bps/s                                ||
#> ||   49%,   0 mins elapsed, rate=5612.0k bps/s                                ||
#> ||   58%,   0 mins elapsed, rate=5641.1k bps/s                                ||
#> ||   66%,   0 mins elapsed, rate=5752.2k bps/s                                ||
#> ||   74%,   0 mins elapsed, rate=5840.8k bps/s                                ||
#> ||   83%,   0 mins elapsed, rate=5906.4k bps/s                                ||
#> ||   91%,   0 mins elapsed, rate=5956.7k bps/s                                ||
#> || 3.1 GB of memory is needed for index building.                             ||
#> || Build the index...                                                         ||
#> ||    8%,   0 mins elapsed, rate=873.2k bps/s                                 ||
#> ||   16%,   0 mins elapsed, rate=1324.9k bps/s                                ||
#> ||   24%,   0 mins elapsed, rate=1565.7k bps/s                                ||
#> ||   33%,   0 mins elapsed, rate=1777.0k bps/s                                ||
#> ||   41%,   0 mins elapsed, rate=1896.5k bps/s                                ||
#> ||   49%,   0 mins elapsed, rate=1989.7k bps/s                                ||
#> ||   58%,   0 mins elapsed, rate=2066.0k bps/s                                ||
#> ||   66%,   0 mins elapsed, rate=2126.6k bps/s                                ||
#> ||   74%,   0 mins elapsed, rate=2194.4k bps/s                                ||
#> ||   83%,   0 mins elapsed, rate=2240.0k bps/s                                ||
#> ||   91%,   0 mins elapsed, rate=2288.5k bps/s                                ||
#> || Save current index block...                                                ||
#> ||  [ 0.0% finished ]                                                         ||
#> ||  [ 10.0% finished ]                                                        ||
#> ||  [ 20.0% finished ]                                                        ||
#> ||  [ 30.0% finished ]                                                        ||
#> ||  [ 40.0% finished ]                                                        ||
#> ||  [ 50.0% finished ]                                                        ||
#> ||  [ 60.0% finished ]                                                        ||
#> ||  [ 70.0% finished ]                                                        ||
#> ||  [ 80.0% finished ]                                                        ||
#> ||  [ 90.0% finished ]                                                        ||
#> ||  [ 100.0% finished ]                                                       ||
#> ||                                                                            ||
#> ||                      Total running time: 0.4 minutes.                      ||
#> ||                    Index target was successfully built.                    ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> [1] "target_1"
readPath <- system.file("extdata", "reads.fastq",package = "MetaScope")
target_map <- align_target(readPath, "target", project_name="subread_target")
#> 
#>         ==========     _____ _    _ ____  _____  ______          _____  
#>         =====         / ____| |  | |  _ \|  __ \|  ____|   /\   |  __ \ 
#>           =====      | (___ | |  | | |_) | |__) | |__     /  \  | |  | |
#>             ====      \___ \| |  | |  _ <|  _  /|  __|   / /\ \ | |  | |
#>               ====    ____) | |__| | |_) | | \ \| |____ / ____ \| |__| |
#>         ==========   |_____/ \____/|____/|_|  \_\______/_/    \_\_____/
#>        Rsubread 2.8.1
#> 
#> //================================= setting ==================================\\
#> ||                                                                            ||
#> || Function      : Read alignment (DNA-Seq)                                   ||
#> || Input file    : reads.fastq                                                ||
#> || Output file   : reads.target.bam (BAM)                                     ||
#> || Index name    : target                                                     ||
#> ||                                                                            ||
#> ||                    ------------------------------------                    ||
#> ||                                                                            ||
#> ||                               Threads : 8                                  ||
#> ||                          Phred offset : 33                                 ||
#> ||                             Min votes : 3 / 10                             ||
#> ||                        Max mismatches : 3                                  ||
#> ||                      Max indel length : 5                                  ||
#> ||            Report multi-mapping reads : yes                                ||
#> || Max alignments per multi-mapping read : 16                                 ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> //================ Running (11-Mar-2022 13:56:00, pid=120161) ================\\
#> ||                                                                            ||
#> || Check the input reads.                                                     ||
#> || The input file contains base space reads.                                  ||
#> || Initialise the memory objects.                                             ||
#> || Estimate the mean read length.                                             ||
#> || WARNING  - The specified Phred score offset (33) seems incorrect.          ||
#> ||            ASCII values of the quality scores of read bases included in    ||
#> ||            the first 1499 reads were found to be within the range of       ||
#> ||            [50,50].                                                        ||
#> ||                                                                            ||
#> || Create the output BAM file.                                                ||
#> || Check the index.                                                           ||
#> || Init the voting space.                                                     ||
#> || Global environment is initialised.                                         ||
#> || Load the 1-th index block...                                               ||
#> || The index block has been loaded.                                           ||
#> || Start read mapping in chunk.                                               ||
#> ||                                                                            ||
#> ||                           Completed successfully.                          ||
#> ||                                                                            ||
#> \\====================================    ====================================//
#> 
#> //================================   Summary =================================\\
#> ||                                                                            ||
#> ||                 Total reads : 1,500                                        ||
#> ||                      Mapped : 903 (60.2%)                                  ||
#> ||             Uniquely mapped : 698                                          ||
#> ||               Multi-mapping : 205                                          ||
#> ||                                                                            ||
#> ||                    Unmapped : 597                                          ||
#> ||                                                                            ||
#> ||                      Indels : 42                                           ||
#> ||                                                                            ||
#> ||                     WARNING : Phred offset (33) incorrect?                 ||
#> ||                                                                            ||
#> ||                Running time : 0.7 minutes                                  ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> Warning: cannot rename file '/scratch/3276553.1.linga/Rtmpp8sDKo/temp_libpath1c89f2e042019/MetaScope/extdata/reads.target.bam' to 'subread_target.bam', reason 'Invalid cross-device link'
#> DONE! Alignments written to subread_target.bam
target_map_sam <- Rsamtools::asSam(target_map, overwrite = TRUE)

## Make and align to multiple reference libraries
mk_subread_index('target.fasta', split = .02)
#> [1] "Library size is 0 GBs. Splitting the library into 2 separate components."
#> [1] "Printed 202 sequences to 2 genome files"
#> 
#>         ==========     _____ _    _ ____  _____  ______          _____  
#>         =====         / ____| |  | |  _ \|  __ \|  ____|   /\   |  __ \ 
#>           =====      | (___ | |  | | |_) | |__) | |__     /  \  | |  | |
#>             ====      \___ \| |  | |  _ <|  _  /|  __|   / /\ \ | |  | |
#>               ====    ____) | |__| | |_) | | \ \| |____ / ____ \| |__| |
#>         ==========   |_____/ \____/|____/|_|  \_\______/_/    \_\_____/
#>        Rsubread 2.8.1
#> 
#> //================================= setting ==================================\\
#> ||                                                                            ||
#> ||                Index name : target_1                                       ||
#> ||               Index space : base space                                     ||
#> ||               Index split : no-split                                       ||
#> ||          Repeat threshold : 100 repeats                                    ||
#> ||              Gapped index : no                                             ||
#> ||                                                                            ||
#> ||       Free / total memory : 317.7GB / 376.4GB                              ||
#> ||                                                                            ||
#> ||               Input files : 1 file in total                                ||
#> ||                             o target_1.fasta                               ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> //================================= Running ==================================\\
#> ||                                                                            ||
#> || Check the integrity of provided reference sequences ...                    ||
#> || No format issues were found                                                ||
#> || Scan uninformative subreads in reference sequences ...                     ||
#> || Estimate the index size...                                                 ||
#> ||    8%,   0 mins elapsed, rate=2752.1k bps/s                                ||
#> ||   16%,   0 mins elapsed, rate=3985.3k bps/s                                ||
#> ||   24%,   0 mins elapsed, rate=4683.7k bps/s                                ||
#> ||   33%,   0 mins elapsed, rate=5129.6k bps/s                                ||
#> ||   49%,   0 mins elapsed, rate=5650.7k bps/s                                ||
#> ||   58%,   0 mins elapsed, rate=5839.6k bps/s                                ||
#> ||   66%,   0 mins elapsed, rate=5984.6k bps/s                                ||
#> ||   74%,   0 mins elapsed, rate=6105.3k bps/s                                ||
#> ||   83%,   0 mins elapsed, rate=6140.5k bps/s                                ||
#> ||   91%,   0 mins elapsed, rate=6184.5k bps/s                                ||
#> || 3.0 GB of memory is needed for index building.                             ||
#> || Build the index...                                                         ||
#> ||    8%,   0 mins elapsed, rate=394.3k bps/s                                 ||
#> ||   16%,   0 mins elapsed, rate=698.5k bps/s                                 ||
#> ||   24%,   0 mins elapsed, rate=939.8k bps/s                                 ||
#> ||   33%,   0 mins elapsed, rate=1136.3k bps/s                                ||
#> ||   49%,   0 mins elapsed, rate=1445.6k bps/s                                ||
#> ||   58%,   0 mins elapsed, rate=1572.0k bps/s                                ||
#> ||   66%,   0 mins elapsed, rate=1676.5k bps/s                                ||
#> ||   74%,   0 mins elapsed, rate=1765.4k bps/s                                ||
#> ||   83%,   0 mins elapsed, rate=1842.7k bps/s                                ||
#> ||   91%,   0 mins elapsed, rate=1911.7k bps/s                                ||
#> || Save current index block...                                                ||
#> ||  [ 0.0% finished ]                                                         ||
#> ||  [ 10.0% finished ]                                                        ||
#> ||  [ 20.0% finished ]                                                        ||
#> ||  [ 30.0% finished ]                                                        ||
#> ||  [ 40.0% finished ]                                                        ||
#> ||  [ 50.0% finished ]                                                        ||
#> ||  [ 60.0% finished ]                                                        ||
#> ||  [ 70.0% finished ]                                                        ||
#> ||  [ 80.0% finished ]                                                        ||
#> ||  [ 90.0% finished ]                                                        ||
#> ||  [ 100.0% finished ]                                                       ||
#> ||                                                                            ||
#> ||                      Total running time: 0.2 minutes.                      ||
#> ||                   Index target_1 was successfully built.                   ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> 
#>         ==========     _____ _    _ ____  _____  ______          _____  
#>         =====         / ____| |  | |  _ \|  __ \|  ____|   /\   |  __ \ 
#>           =====      | (___ | |  | | |_) | |__) | |__     /  \  | |  | |
#>             ====      \___ \| |  | |  _ <|  _  /|  __|   / /\ \ | |  | |
#>               ====    ____) | |__| | |_) | | \ \| |____ / ____ \| |__| |
#>         ==========   |_____/ \____/|____/|_|  \_\______/_/    \_\_____/
#>        Rsubread 2.8.1
#> 
#> //================================= setting ==================================\\
#> ||                                                                            ||
#> ||                Index name : target_2                                       ||
#> ||               Index space : base space                                     ||
#> ||               Index split : no-split                                       ||
#> ||          Repeat threshold : 100 repeats                                    ||
#> ||              Gapped index : no                                             ||
#> ||                                                                            ||
#> ||       Free / total memory : 317.9GB / 376.4GB                              ||
#> ||                                                                            ||
#> ||               Input files : 1 file in total                                ||
#> ||                             o target_2.fasta                               ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> //================================= Running ==================================\\
#> ||                                                                            ||
#> || Check the integrity of provided reference sequences ...                    ||
#> || No format issues were found                                                ||
#> || Scan uninformative subreads in reference sequences ...                     ||
#> || 2 uninformative subreads were found.                                       ||
#> || These subreads were excluded from index building.                          ||
#> || Estimate the index size...                                                 ||
#> ||    8%,   0 mins elapsed, rate=3946.9k bps/s                                ||
#> ||   16%,   0 mins elapsed, rate=5062.1k bps/s                                ||
#> ||   33%,   0 mins elapsed, rate=5907.8k bps/s                                ||
#> ||   41%,   0 mins elapsed, rate=6108.0k bps/s                                ||
#> ||   49%,   0 mins elapsed, rate=6250.1k bps/s                                ||
#> ||   58%,   0 mins elapsed, rate=6354.3k bps/s                                ||
#> ||   66%,   0 mins elapsed, rate=6434.3k bps/s                                ||
#> ||   74%,   0 mins elapsed, rate=6490.1k bps/s                                ||
#> ||   83%,   0 mins elapsed, rate=6543.1k bps/s                                ||
#> ||   91%,   0 mins elapsed, rate=6587.8k bps/s                                ||
#> || 3.1 GB of memory is needed for index building.                             ||
#> || Build the index...                                                         ||
#> ||    8%,   0 mins elapsed, rate=687.0k bps/s                                 ||
#> ||   16%,   0 mins elapsed, rate=1111.4k bps/s                                ||
#> ||   33%,   0 mins elapsed, rate=1620.5k bps/s                                ||
#> ||   41%,   0 mins elapsed, rate=1778.7k bps/s                                ||
#> ||   49%,   0 mins elapsed, rate=1902.8k bps/s                                ||
#> ||   58%,   0 mins elapsed, rate=2001.6k bps/s                                ||
#> ||   66%,   0 mins elapsed, rate=2082.7k bps/s                                ||
#> ||   74%,   0 mins elapsed, rate=2150.6k bps/s                                ||
#> ||   83%,   0 mins elapsed, rate=2211.4k bps/s                                ||
#> ||   91%,   0 mins elapsed, rate=2260.4k bps/s                                ||
#> || Save current index block...                                                ||
#> ||  [ 0.0% finished ]                                                         ||
#> ||  [ 10.0% finished ]                                                        ||
#> ||  [ 20.0% finished ]                                                        ||
#> ||  [ 30.0% finished ]                                                        ||
#> ||  [ 40.0% finished ]                                                        ||
#> ||  [ 50.0% finished ]                                                        ||
#> ||  [ 60.0% finished ]                                                        ||
#> ||  [ 70.0% finished ]                                                        ||
#> ||  [ 80.0% finished ]                                                        ||
#> ||  [ 90.0% finished ]                                                        ||
#> ||  [ 100.0% finished ]                                                       ||
#> ||                                                                            ||
#> ||                      Total running time: 0.3 minutes.                      ||
#> ||                   Index target_2 was successfully built.                   ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> [1] "target_1" "target_2"
targLibs <- c("target_1", "target_2")
readPath <- system.file("extdata", "reads.fastq",package = "MetaScope")
target_map <- align_target(readPath, targLibs, project_name = "subread_target")
#> 
#>         ==========     _____ _    _ ____  _____  ______          _____  
#>         =====         / ____| |  | |  _ \|  __ \|  ____|   /\   |  __ \ 
#>           =====      | (___ | |  | | |_) | |__) | |__     /  \  | |  | |
#>             ====      \___ \| |  | |  _ <|  _  /|  __|   / /\ \ | |  | |
#>               ====    ____) | |__| | |_) | | \ \| |____ / ____ \| |__| |
#>         ==========   |_____/ \____/|____/|_|  \_\______/_/    \_\_____/
#>        Rsubread 2.8.1
#> 
#> //================================= setting ==================================\\
#> ||                                                                            ||
#> || Function      : Read alignment (DNA-Seq)                                   ||
#> || Input file    : reads.fastq                                                ||
#> || Output file   : reads.target_1.bam (BAM)                                   ||
#> || Index name    : target_1                                                   ||
#> ||                                                                            ||
#> ||                    ------------------------------------                    ||
#> ||                                                                            ||
#> ||                               Threads : 8                                  ||
#> ||                          Phred offset : 33                                 ||
#> ||                             Min votes : 3 / 10                             ||
#> ||                        Max mismatches : 3                                  ||
#> ||                      Max indel length : 5                                  ||
#> ||            Report multi-mapping reads : yes                                ||
#> || Max alignments per multi-mapping read : 16                                 ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> //================ Running (11-Mar-2022 13:57:19, pid=120161) ================\\
#> ||                                                                            ||
#> || Check the input reads.                                                     ||
#> || The input file contains base space reads.                                  ||
#> || Initialise the memory objects.                                             ||
#> || Estimate the mean read length.                                             ||
#> || WARNING  - The specified Phred score offset (33) seems incorrect.          ||
#> ||            ASCII values of the quality scores of read bases included in    ||
#> ||            the first 1499 reads were found to be within the range of       ||
#> ||            [50,50].                                                        ||
#> ||                                                                            ||
#> || Create the output BAM file.                                                ||
#> || Check the index.                                                           ||
#> || Init the voting space.                                                     ||
#> || Global environment is initialised.                                         ||
#> || Load the 1-th index block...                                               ||
#> || The index block has been loaded.                                           ||
#> || Start read mapping in chunk.                                               ||
#> ||                                                                            ||
#> ||                           Completed successfully.                          ||
#> ||                                                                            ||
#> \\====================================    ====================================//
#> 
#> //================================   Summary =================================\\
#> ||                                                                            ||
#> ||                 Total reads : 1,500                                        ||
#> ||                      Mapped : 892 (59.5%)                                  ||
#> ||             Uniquely mapped : 750                                          ||
#> ||               Multi-mapping : 142                                          ||
#> ||                                                                            ||
#> ||                    Unmapped : 608                                          ||
#> ||                                                                            ||
#> ||                      Indels : 31                                           ||
#> ||                                                                            ||
#> ||                     WARNING : Phred offset (33) incorrect?                 ||
#> ||                                                                            ||
#> ||                Running time : 0.7 minutes                                  ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> 
#>         ==========     _____ _    _ ____  _____  ______          _____  
#>         =====         / ____| |  | |  _ \|  __ \|  ____|   /\   |  __ \ 
#>           =====      | (___ | |  | | |_) | |__) | |__     /  \  | |  | |
#>             ====      \___ \| |  | |  _ <|  _  /|  __|   / /\ \ | |  | |
#>               ====    ____) | |__| | |_) | | \ \| |____ / ____ \| |__| |
#>         ==========   |_____/ \____/|____/|_|  \_\______/_/    \_\_____/
#>        Rsubread 2.8.1
#> 
#> //================================= setting ==================================\\
#> ||                                                                            ||
#> || Function      : Read alignment (DNA-Seq)                                   ||
#> || Input file    : reads.fastq                                                ||
#> || Output file   : reads.target_2.bam (BAM)                                   ||
#> || Index name    : target_2                                                   ||
#> ||                                                                            ||
#> ||                    ------------------------------------                    ||
#> ||                                                                            ||
#> ||                               Threads : 8                                  ||
#> ||                          Phred offset : 33                                 ||
#> ||                             Min votes : 3 / 10                             ||
#> ||                        Max mismatches : 3                                  ||
#> ||                      Max indel length : 5                                  ||
#> ||            Report multi-mapping reads : yes                                ||
#> || Max alignments per multi-mapping read : 16                                 ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> //================ Running (11-Mar-2022 13:57:59, pid=120161) ================\\
#> ||                                                                            ||
#> || Check the input reads.                                                     ||
#> || The input file contains base space reads.                                  ||
#> || Initialise the memory objects.                                             ||
#> || Estimate the mean read length.                                             ||
#> || WARNING  - The specified Phred score offset (33) seems incorrect.          ||
#> ||            ASCII values of the quality scores of read bases included in    ||
#> ||            the first 1499 reads were found to be within the range of       ||
#> ||            [50,50].                                                        ||
#> ||                                                                            ||
#> || Create the output BAM file.                                                ||
#> || Check the index.                                                           ||
#> || Init the voting space.                                                     ||
#> || Global environment is initialised.                                         ||
#> || Load the 1-th index block...                                               ||
#> || The index block has been loaded.                                           ||
#> || Start read mapping in chunk.                                               ||
#> ||                                                                            ||
#> ||                           Completed successfully.                          ||
#> ||                                                                            ||
#> \\====================================    ====================================//
#> 
#> //================================   Summary =================================\\
#> ||                                                                            ||
#> ||                 Total reads : 1,500                                        ||
#> ||                      Mapped : 599 (39.9%)                                  ||
#> ||             Uniquely mapped : 140                                          ||
#> ||               Multi-mapping : 459                                          ||
#> ||                                                                            ||
#> ||                    Unmapped : 901                                          ||
#> ||                                                                            ||
#> ||                      Indels : 108                                          ||
#> ||                                                                            ||
#> ||                     WARNING : Phred offset (33) incorrect?                 ||
#> ||                                                                            ||
#> ||                Running time : 0.6 minutes                                  ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> [1] "Merging the bam files into subread_target.bam"
#> [1] "Making a combined header file: subread_target_header.sam"
#> [1] "Merging and sorting .bam files"
#> DONE! Alignments written to subread_target.bam