A universal parameter object for Bowtie 2 metagenomic alignment where the host genome is thought to be absent from the reference database

This character string provides several Bowtie 2 options to conduct an alignment useful for metagenomes, especially in the case where a genome may not be present in the reference database. This object can be used with functions that use the Bowtie 2 aligner through the Rbowtie2 package, namely align_target_bowtie() and filter_host_bowtie. These settings can be substituted for default settings by passing to the bowtie2_options argument.

Usage

data(bt2_missing_params)

Format

list

Details

The default parameters listed in this object are "–local -R 2 -N 0 -L 25 -i S,1,0.75 -k 5 –score-min L,0,1.4".

Further delineation of Bowtie 2 parameters is provided in the Bowtie 2 manual.

Examples

data("bt2_missing_params")