Helper function for demultiplexing

Helper function for demultiplexing sequencing reads, designed in a way to allow for parallelization accross barcodes (parallel extraction of reads by barcode). This function takes a specific barcode (numeric index) from lists of sample names/barcodes, a Biostrings::DNAStringSet of barcodes by sequence header, and a Biostrings::QualityScaledXStringSet of reads corresponding to the barcodes. Based on the barcode index given, it extracts all reads for the indexed barcode and writes all the reads from that barcode to a separate .fastq file.

Usage

extractReads(
  barcodeIndex,
  barcodes,
  sampleNames,
  index,
  reads,
  location = "./demultiplex_fastq",
  rcBarcodes = TRUE,
  hDist = 0
)

Arguments

barcodeIndex

Which barcode (integer number or index) in the barcodes or sample name to use for read extraction

barcodes

A list of all barcodes in the sequencing dataset. Correlates and in same order as sampleNames

sampleNames

A list of sample names or identifiers associated with each barcode in the barcodes list

index

A Biostrings::DNAStringSet that contains the read headers and barcode sequence for each header in the sequence slot

reads

A Biostrings::QualityScaledXStringSet that has the same headers and order as the index file, but contains the read sequences and their quality scores

location

A directory location to store the demuliplexed read files. Defaults to generate a new subdirectory at './demultiplex_fastq'

rcBarcodes

Should the barcode indexes in the barcodes list be reverse complemented to match the sequences in the index DNAStringSet? Defaults to TRUE

hDist

Uses a Hamming Distance or number of base differences to allow for inexact matches for the barcodes/indexes. Defaults to 0. Warning: if the Hamming Distance is >=1 and this leads to inexact index matches to more than one barcode, that read will be written to more than one demultiplexed read files

Value

Writes a single .fastq file that contains all reads whose index matches the barcode specified. This file will be written to the location directory, and will be named based on the specified sampleName and barcode, e.g. './demultiplex_fastq/SampleName1_GGAATTATCGGT.fastq.gz'

Examples

## Load example barcode, index, and read data into R session:
barcodePath <- system.file("extdata", "barcodes.txt", package = "MetaScope")
bcFile <- read.table(barcodePath, sep = "\t", header = TRUE)

indexPath <- system.file("extdata", "virus_example_index.fastq",
package = "MetaScope")
inds <- Biostrings::readDNAStringSet(indexPath, format = "fastq")

readPath <- system.file("extdata", "virus_example.fastq",
                        package = "MetaScope")
reads <- Biostrings::readQualityScaledDNAStringSet(readPath)

## Extract reads from the first barcode
results <- extractReads(1, bcFile[, 2], bcFile[, 1], inds, reads,
                        rcBarcodes = FALSE, location = ".")
#> Finding reads for barcode: TCCACGT
#> 	Found 25 reads, writing reads to: ./CDV_TCCACGT.fastq.gz
results
#> $output_file
#> [1] "./CDV_TCCACGT.fastq.gz"
#> 
#> $numberOfReads
#> [1] 25
#> 
#> $matchedIndexes
#>                                                              [,1]
#> K00153:13:H2WGGBBXX:7:1101:13149:1000 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:22711:1000 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:4593:1017 1:N:0:TCCACGT+TTGCCAC   TRUE
#> K00153:13:H2WGGBBXX:7:1101:6115:1017 1:N:0:TCCACGT+TTGCCAC   TRUE
#> K00153:13:H2WGGBBXX:7:1101:11606:1035 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:14692:1035 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:15869:1035 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:21635:1035 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:7658:1053 1:N:0:TCCACGT+TTGCCAC   TRUE
#> K00153:13:H2WGGBBXX:7:1101:12611:1053 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:13261:1053 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:14012:1088 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:15433:1088 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:11363:1105 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:15321:1105 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:8328:1123 1:N:0:TCCACGT+TTGCCAC   TRUE
#> K00153:13:H2WGGBBXX:7:1101:16022:1123 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:19249:1123 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:18569:1141 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:13271:1176 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:19270:1193 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:11322:1211 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:12459:1211 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:13677:1211 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:18245:1211 1:N:0:TCCACGT+TTGCCAC  TRUE
#> K00153:13:H2WGGBBXX:7:1101:8764:1000 1:N:0:ACAGGCT+GTCTTCT  FALSE
#> K00153:13:H2WGGBBXX:7:1101:10246:1000 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:10510:1000 1:N:0:ACAGGCG+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:13190:1000 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:13860:1000 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:13941:1000 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:14002:1000 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:14388:1000 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:15301:1000 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:15809:1000 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:16052:1000 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:16235:1000 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:17534:1000 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:18752:1000 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:22447:1000 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:22873:1000 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:23502:1000 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:9404:1017 1:N:0:ACAGGCT+GTCTTCT  FALSE
#> K00153:13:H2WGGBBXX:7:1101:11901:1017 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:13890:1017 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:14134:1017 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:15169:1017 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:15636:1017 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:16590:1017 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:16773:1017 1:N:0:ACAGGCT+GTCTTCT FALSE
#> K00153:13:H2WGGBBXX:7:1101:13718:1000 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:16722:1000 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:8409:1017 1:N:0:ATCGTGC+TTCGTGT  FALSE
#> K00153:13:H2WGGBBXX:7:1101:10155:1017 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:13220:1017 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:15474:1017 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:21218:1017 1:N:0:ATCGTGC+TTCGGGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:11424:1035 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:15524:1035 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:19645:1035 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:9850:1053 1:N:0:ATCGTGC+TTCGTGT  FALSE
#> K00153:13:H2WGGBBXX:7:1101:17118:1053 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:10855:1070 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:12855:1088 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:18113:1088 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:22051:1088 1:N:0:ATCTTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:26090:1088 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:7140:1105 1:N:0:ATCGTGC+TTCGTGT  FALSE
#> K00153:13:H2WGGBBXX:7:1101:11941:1123 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:13200:1123 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:13870:1123 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:19534:1123 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:10673:1141 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:13839:1141 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:14448:1141 1:N:0:ATCGTGC+TTCGTGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:18042:1000 1:N:0:ACTACAG+TCGCAGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:20762:1000 1:N:0:ACTACAG+ACGCAGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:26050:1017 1:N:0:ACTACAG+TCGCAGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:5334:1035 1:N:0:ACTACAG+TCGCAGT  FALSE
#> K00153:13:H2WGGBBXX:7:1101:7181:1035 1:N:0:ACTACAG+TCGCAGT  FALSE
#> K00153:13:H2WGGBBXX:7:1101:12155:1035 1:N:0:ACTACAG+TCGCAGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:12946:1035 1:N:0:ACTACAG+TCGCAGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:22447:1035 1:N:0:ACTACAG+TCGCAGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:9282:1053 1:N:0:ACTACAG+TCGCAGT  FALSE
#> K00153:13:H2WGGBBXX:7:1101:9668:1053 1:N:0:ACTACAG+TCGCAGT  FALSE
#> K00153:13:H2WGGBBXX:7:1101:12835:1053 1:N:0:ACTACAG+TCGCAGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:15311:1053 1:N:0:ACTACAG+TCGCAGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:15453:1053 1:N:0:ACTACAG+TCGCAGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:7668:1070 1:N:0:ACTACAG+TCGCAGT  FALSE
#> K00153:13:H2WGGBBXX:7:1101:18326:1070 1:N:0:ACTACAG+TCGCAGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:22873:1070 1:N:0:ACTACAG+TCGCAGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:13606:1088 1:N:0:ACTACAG+TCGCAGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:16407:1088 1:N:0:ACTACAG+TCGCAGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:19554:1088 1:N:0:ACTACAG+TCGCAGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:9110:1105 1:N:0:ACTACAG+TCGCAGT  FALSE
#> K00153:13:H2WGGBBXX:7:1101:12094:1105 1:N:0:ACTACAG+TCGCAGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:13799:1105 1:N:0:ACTACAG+TCGCAGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:8572:1123 1:N:0:ACTACAG+TCGCAGT  FALSE
#> K00153:13:H2WGGBBXX:7:1101:18904:1123 1:N:0:ACTACAG+TCGCAGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:21320:1123 1:N:0:ACTACAG+TCGCAGT FALSE
#> K00153:13:H2WGGBBXX:7:1101:7628:1000 1:N:0:AAGTCGC+TGCTTGA  FALSE
#> K00153:13:H2WGGBBXX:7:1101:16397:1000 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:18285:1000 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:21310:1000 1:N:0:AAGTCGC+TGCTAGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:24355:1000 1:N:0:AAGTCGC+GGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:20203:1017 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:19503:1035 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:12956:1053 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:14611:1070 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:20599:1070 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:9911:1088 1:N:0:AAGTCGC+TGCTTGA  FALSE
#> K00153:13:H2WGGBBXX:7:1101:20650:1088 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:12479:1105 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:12703:1105 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:17230:1105 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:15595:1123 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:19463:1141 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:30709:1141 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:11251:1158 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:11454:1158 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:19107:1158 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:6887:1193 1:N:0:AAGTCGC+TGCTTGA  FALSE
#> K00153:13:H2WGGBBXX:7:1101:20955:1193 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:14388:1211 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:15859:1228 1:N:0:AAGTCGC+TGCTTGA FALSE
#> K00153:13:H2WGGBBXX:7:1101:15849:1000 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:8247:1017 1:N:0:TCTCAGG+GTACCCA  FALSE
#> K00153:13:H2WGGBBXX:7:1101:9526:1017 1:N:0:TCTCAGG+GTACTCA  FALSE
#> K00153:13:H2WGGBBXX:7:1101:11840:1017 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:14398:1017 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:18265:1035 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:17219:1053 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:18985:1053 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:20203:1053 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:15494:1088 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:16742:1105 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:17737:1105 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:20092:1105 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:20193:1105 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:13565:1123 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:10043:1141 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:14164:1141 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:17859:1141 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:13200:1158 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:6328:1176 1:N:0:TCTCAGG+GTACTCA  FALSE
#> K00153:13:H2WGGBBXX:7:1101:11241:1176 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:14002:1176 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:15220:1176 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:18590:1176 1:N:0:TCTCAGG+GTACTCA FALSE
#> K00153:13:H2WGGBBXX:7:1101:16529:1193 1:N:0:TCTCAGG+GTACTCA FALSE
#> 

## Extract reads from multiple barcodes
more_results <- lapply(1:6, extractReads, bcFile[, 2], bcFile[, 1], inds,
                       reads, rcBarcodes = FALSE, location = ".")
#> Finding reads for barcode: TCCACGT
#> 	Found 25 reads, writing reads to: ./CDV_TCCACGT.fastq.gz
#> Finding reads for barcode: ACAGGCT
#> 	Found 25 reads, writing reads to: ./LaCrosse_ACAGGCT.fastq.gz
#> Finding reads for barcode: ATCGTGC
#> 	Found 25 reads, writing reads to: ./RSV_ATCGTGC.fastq.gz
#> Finding reads for barcode: ACTACAG
#> 	Found 25 reads, writing reads to: ./EboV_ACTACAG.fastq.gz
#> Finding reads for barcode: AAGTCGC
#> 	Found 25 reads, writing reads to: ./Measles_AAGTCGC.fastq.gz
#> Finding reads for barcode: TCTCAGG
#> 	Found 25 reads, writing reads to: ./VSV_TCTCAGG.fastq.gz

## BiocParallel application
multicoreParam <- BiocParallel::MulticoreParam(workers = 3)
parallel_results <- BiocParallel::bplapply(1:6, extractReads, bcFile[, 2], 
                                           bcFile[, 1], inds, reads, 
                                           rcBarcodes = FALSE, 
                                           location = ".", 
                                           BPPARAM = multicoreParam)