Helper Function for MetaScope ID

Used to create plots of genome coverage for any number of accession numbers

Usage

locations(
  which_taxid,
  which_genome,
  accessions,
  taxids,
  reads,
  out_base,
  out_dir
)

Arguments

which_taxid

Which taxid to plot

which_genome

Which genome to plot

accessions

List of accessions from metascope_id()

taxids

List of accessions from metascope_id()

reads

List of reads from input file

out_base

The basename of the input file

out_dir

The path to the input file

Value

A plot of the read coverage for a given genome