Make a Subread index

This function is a wrapper for the Rsubread::buildindex function. It will generate one or more Subread indexes from a .fasta file. If the library is too large (default >4GB) it will automatically be split into multiple indexes, with _1, _2, etc at the end of the ref_lib basename.

Usage

mk_subread_index(ref_lib, split = 4, mem = 8000)

Arguments

ref_lib

The name/location of the reference library file, in (uncompressed) .fasta format

split

The maximum allowed size of the genome file (in GB). If the ref_lib file is larger than this, the function will split the library into multiple parts

mem

The maximum amount of memory (in MB) that can be used by the index generation process (used by the Rsubread::buildindex function)

Value

Creates one or more Subread indexes for the supplied reference .fasta file. If multiple indexes are created, the libraries will be named the ref_lib basename plus _1, _2, etc. The function returns the names of the folders holding these files.

Examples

#### Create a subread index from the example reference library

## Create object with path to example reference library
refPath <- system.file("extdata","target.fasta", package = "MetaScope")

## Copy the example reference library to the current directory
file.copy(from = refPath, to = file.path(".", "target.fasta"))
#> [1] FALSE

## Make subread index of reference library
mk_subread_index('target.fasta')
#> [1] "Library size is 0 GBs. Building the Rsubread index"
#> 
#>         ==========     _____ _    _ ____  _____  ______          _____  
#>         =====         / ____| |  | |  _ \|  __ \|  ____|   /\   |  __ \ 
#>           =====      | (___ | |  | | |_) | |__) | |__     /  \  | |  | |
#>             ====      \___ \| |  | |  _ <|  _  /|  __|   / /\ \ | |  | |
#>               ====    ____) | |__| | |_) | | \ \| |____ / ____ \| |__| |
#>         ==========   |_____/ \____/|____/|_|  \_\______/_/    \_\_____/
#>        Rsubread 2.8.1
#> 
#> //================================= setting ==================================\\
#> ||                                                                            ||
#> ||                Index name : target                                         ||
#> ||               Index space : base space                                     ||
#> ||               Index split : no-split                                       ||
#> ||          Repeat threshold : 100 repeats                                    ||
#> ||              Gapped index : no                                             ||
#> ||                                                                            ||
#> ||       Free / total memory : 343.6GB / 376.4GB                              ||
#> ||                                                                            ||
#> ||               Input files : 1 file in total                                ||
#> ||                             o target.fasta                                 ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> //================================= Running ==================================\\
#> ||                                                                            ||
#> || Check the integrity of provided reference sequences ...                    ||
#> || No format issues were found                                                ||
#> || Scan uninformative subreads in reference sequences ...                     ||
#> || 45 uninformative subreads were found.                                      ||
#> || These subreads were excluded from index building.                          ||
#> || Estimate the index size...                                                 ||
#> ||    8%,   0 mins elapsed, rate=4643.9k bps/s                                ||
#> ||   16%,   0 mins elapsed, rate=5563.2k bps/s                                ||
#> ||   24%,   0 mins elapsed, rate=5955.0k bps/s                                ||
#> ||   33%,   0 mins elapsed, rate=6168.4k bps/s                                ||
#> ||   41%,   0 mins elapsed, rate=6336.5k bps/s                                ||
#> ||   49%,   0 mins elapsed, rate=6429.4k bps/s                                ||
#> ||   58%,   0 mins elapsed, rate=6498.3k bps/s                                ||
#> ||   66%,   0 mins elapsed, rate=6550.3k bps/s                                ||
#> ||   74%,   0 mins elapsed, rate=6589.4k bps/s                                ||
#> ||   83%,   0 mins elapsed, rate=6623.8k bps/s                                ||
#> ||   91%,   0 mins elapsed, rate=6652.3k bps/s                                ||
#> || 3.1 GB of memory is needed for index building.                             ||
#> || Build the index...                                                         ||
#> ||    8%,   0 mins elapsed, rate=957.3k bps/s                                 ||
#> ||   16%,   0 mins elapsed, rate=1452.0k bps/s                                ||
#> ||   24%,   0 mins elapsed, rate=1752.8k bps/s                                ||
#> ||   33%,   0 mins elapsed, rate=1959.7k bps/s                                ||
#> ||   41%,   0 mins elapsed, rate=2116.3k bps/s                                ||
#> ||   49%,   0 mins elapsed, rate=2225.2k bps/s                                ||
#> ||   58%,   0 mins elapsed, rate=2310.0k bps/s                                ||
#> ||   66%,   0 mins elapsed, rate=2377.8k bps/s                                ||
#> ||   74%,   0 mins elapsed, rate=2438.4k bps/s                                ||
#> ||   83%,   0 mins elapsed, rate=2485.5k bps/s                                ||
#> ||   91%,   0 mins elapsed, rate=2524.1k bps/s                                ||
#> || Save current index block...                                                ||
#> ||  [ 0.0% finished ]                                                         ||
#> ||  [ 10.0% finished ]                                                        ||
#> ||  [ 20.0% finished ]                                                        ||
#> ||  [ 30.0% finished ]                                                        ||
#> ||  [ 40.0% finished ]                                                        ||
#> ||  [ 50.0% finished ]                                                        ||
#> ||  [ 60.0% finished ]                                                        ||
#> ||  [ 70.0% finished ]                                                        ||
#> ||  [ 80.0% finished ]                                                        ||
#> ||  [ 90.0% finished ]                                                        ||
#> ||  [ 100.0% finished ]                                                       ||
#> ||                                                                            ||
#> ||                      Total running time: 0.5 minutes.                      ||
#> ||                    Index target was successfully built.                    ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> [1] "target_1"

#### Create multiple subread indexes from the example reference library

## Create object with path to example reference library
refPath <- system.file("extdata","target.fasta", package = "MetaScope")

## Copy the example reference library to the current directory
file.copy(from = refPath, to = file.path(".", "target.fasta"))
#> [1] FALSE

## Make multiple subread indexes of reference library
mk_subread_index('target.fasta', split = .02)
#> [1] "Library size is 0 GBs. Splitting the library into 2 separate components."
#> [1] "Printed 202 sequences to 2 genome files"
#> 
#>         ==========     _____ _    _ ____  _____  ______          _____  
#>         =====         / ____| |  | |  _ \|  __ \|  ____|   /\   |  __ \ 
#>           =====      | (___ | |  | | |_) | |__) | |__     /  \  | |  | |
#>             ====      \___ \| |  | |  _ <|  _  /|  __|   / /\ \ | |  | |
#>               ====    ____) | |__| | |_) | | \ \| |____ / ____ \| |__| |
#>         ==========   |_____/ \____/|____/|_|  \_\______/_/    \_\_____/
#>        Rsubread 2.8.1
#> 
#> //================================= setting ==================================\\
#> ||                                                                            ||
#> ||                Index name : target_1                                       ||
#> ||               Index space : base space                                     ||
#> ||               Index split : no-split                                       ||
#> ||          Repeat threshold : 100 repeats                                    ||
#> ||              Gapped index : no                                             ||
#> ||                                                                            ||
#> ||       Free / total memory : 343.6GB / 376.4GB                              ||
#> ||                                                                            ||
#> ||               Input files : 1 file in total                                ||
#> ||                             o target_1.fasta                               ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> //================================= Running ==================================\\
#> ||                                                                            ||
#> || Check the integrity of provided reference sequences ...                    ||
#> || No format issues were found                                                ||
#> || Scan uninformative subreads in reference sequences ...                     ||
#> || Estimate the index size...                                                 ||
#> ||    8%,   0 mins elapsed, rate=2775.9k bps/s                                ||
#> ||   16%,   0 mins elapsed, rate=3950.8k bps/s                                ||
#> ||   24%,   0 mins elapsed, rate=4599.2k bps/s                                ||
#> ||   33%,   0 mins elapsed, rate=5004.8k bps/s                                ||
#> ||   49%,   0 mins elapsed, rate=5505.7k bps/s                                ||
#> ||   58%,   0 mins elapsed, rate=5688.3k bps/s                                ||
#> ||   66%,   0 mins elapsed, rate=5816.9k bps/s                                ||
#> ||   74%,   0 mins elapsed, rate=5919.1k bps/s                                ||
#> ||   83%,   0 mins elapsed, rate=6001.3k bps/s                                ||
#> ||   91%,   0 mins elapsed, rate=6070.3k bps/s                                ||
#> || 3.0 GB of memory is needed for index building.                             ||
#> || Build the index...                                                         ||
#> ||    8%,   0 mins elapsed, rate=407.5k bps/s                                 ||
#> ||   16%,   0 mins elapsed, rate=716.3k bps/s                                 ||
#> ||   24%,   0 mins elapsed, rate=960.5k bps/s                                 ||
#> ||   33%,   0 mins elapsed, rate=1158.1k bps/s                                ||
#> ||   49%,   0 mins elapsed, rate=1468.8k bps/s                                ||
#> ||   58%,   0 mins elapsed, rate=1594.2k bps/s                                ||
#> ||   66%,   0 mins elapsed, rate=1688.6k bps/s                                ||
#> ||   74%,   0 mins elapsed, rate=1773.6k bps/s                                ||
#> ||   83%,   0 mins elapsed, rate=1849.5k bps/s                                ||
#> ||   91%,   0 mins elapsed, rate=1917.2k bps/s                                ||
#> || Save current index block...                                                ||
#> ||  [ 0.0% finished ]                                                         ||
#> ||  [ 10.0% finished ]                                                        ||
#> ||  [ 20.0% finished ]                                                        ||
#> ||  [ 30.0% finished ]                                                        ||
#> ||  [ 40.0% finished ]                                                        ||
#> ||  [ 50.0% finished ]                                                        ||
#> ||  [ 60.0% finished ]                                                        ||
#> ||  [ 70.0% finished ]                                                        ||
#> ||  [ 80.0% finished ]                                                        ||
#> ||  [ 90.0% finished ]                                                        ||
#> ||  [ 100.0% finished ]                                                       ||
#> ||                                                                            ||
#> ||                      Total running time: 0.3 minutes.                      ||
#> ||                   Index target_1 was successfully built.                   ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> 
#>         ==========     _____ _    _ ____  _____  ______          _____  
#>         =====         / ____| |  | |  _ \|  __ \|  ____|   /\   |  __ \ 
#>           =====      | (___ | |  | | |_) | |__) | |__     /  \  | |  | |
#>             ====      \___ \| |  | |  _ <|  _  /|  __|   / /\ \ | |  | |
#>               ====    ____) | |__| | |_) | | \ \| |____ / ____ \| |__| |
#>         ==========   |_____/ \____/|____/|_|  \_\______/_/    \_\_____/
#>        Rsubread 2.8.1
#> 
#> //================================= setting ==================================\\
#> ||                                                                            ||
#> ||                Index name : target_2                                       ||
#> ||               Index space : base space                                     ||
#> ||               Index split : no-split                                       ||
#> ||          Repeat threshold : 100 repeats                                    ||
#> ||              Gapped index : no                                             ||
#> ||                                                                            ||
#> ||       Free / total memory : 343.5GB / 376.4GB                              ||
#> ||                                                                            ||
#> ||               Input files : 1 file in total                                ||
#> ||                             o target_2.fasta                               ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> //================================= Running ==================================\\
#> ||                                                                            ||
#> || Check the integrity of provided reference sequences ...                    ||
#> || No format issues were found                                                ||
#> || Scan uninformative subreads in reference sequences ...                     ||
#> || 2 uninformative subreads were found.                                       ||
#> || These subreads were excluded from index building.                          ||
#> || Estimate the index size...                                                 ||
#> ||    8%,   0 mins elapsed, rate=3977.1k bps/s                                ||
#> ||   16%,   0 mins elapsed, rate=5032.2k bps/s                                ||
#> ||   33%,   0 mins elapsed, rate=5749.3k bps/s                                ||
#> ||   41%,   0 mins elapsed, rate=5943.6k bps/s                                ||
#> ||   49%,   0 mins elapsed, rate=6080.0k bps/s                                ||
#> ||   58%,   0 mins elapsed, rate=6180.9k bps/s                                ||
#> ||   66%,   0 mins elapsed, rate=6260.2k bps/s                                ||
#> ||   74%,   0 mins elapsed, rate=6315.1k bps/s                                ||
#> ||   83%,   0 mins elapsed, rate=6366.7k bps/s                                ||
#> ||   91%,   0 mins elapsed, rate=6409.6k bps/s                                ||
#> || 3.1 GB of memory is needed for index building.                             ||
#> || Build the index...                                                         ||
#> ||    8%,   0 mins elapsed, rate=719.0k bps/s                                 ||
#> ||   16%,   0 mins elapsed, rate=1162.9k bps/s                                ||
#> ||   33%,   0 mins elapsed, rate=1693.7k bps/s                                ||
#> ||   41%,   0 mins elapsed, rate=1858.1k bps/s                                ||
#> ||   49%,   0 mins elapsed, rate=1986.5k bps/s                                ||
#> ||   58%,   0 mins elapsed, rate=2089.7k bps/s                                ||
#> ||   66%,   0 mins elapsed, rate=2173.2k bps/s                                ||
#> ||   74%,   0 mins elapsed, rate=2243.9k bps/s                                ||
#> ||   83%,   0 mins elapsed, rate=2307.4k bps/s                                ||
#> ||   91%,   0 mins elapsed, rate=2358.7k bps/s                                ||
#> || Save current index block...                                                ||
#> ||  [ 0.0% finished ]                                                         ||
#> ||  [ 10.0% finished ]                                                        ||
#> ||  [ 20.0% finished ]                                                        ||
#> ||  [ 30.0% finished ]                                                        ||
#> ||  [ 40.0% finished ]                                                        ||
#> ||  [ 50.0% finished ]                                                        ||
#> ||  [ 60.0% finished ]                                                        ||
#> ||  [ 70.0% finished ]                                                        ||
#> ||  [ 80.0% finished ]                                                        ||
#> ||  [ 90.0% finished ]                                                        ||
#> ||  [ 100.0% finished ]                                                       ||
#> ||                                                                            ||
#> ||                      Total running time: 0.4 minutes.                      ||
#> ||                   Index target_2 was successfully built.                   ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> [1] "target_1" "target_2"