Align microbiome reads to a set of reference libraries

This is the main MetaScope target library mapping function, using Rsubread and multiple libraries. Aligns to each library separately, filters unmapped reads from each file, and then merges and sorts the .bam files from each library into one output file. If desired, output can be passed to `filter_host()` to remove reads that also map to filter library genomes.

Usage

align_target(
  read1,
  read2 = NULL,
  lib_dir = NULL,
  libs,
  threads = 1,
  align_file = tools::file_path_sans_ext(read1),
  subread_options = align_details
)

Arguments

read1

Path to the .fastq file to align.

read2

Optional: Location of the mate pair .fastq file to align.

lib_dir

Path to the index files for all libraries.

libs

A vector of character strings giving the basenames of the Subread index files for alignment. If ALL indices to be used are located in the current working directory, set lib_dir = NULL. Default is lib_dir = NULL.

threads

The number of threads that can be utilized by the function. Default is 1 thread.

align_file

The basename of the output alignment file (without trailing .bam extension).

subread_options

A named list specifying alignment parameters for the Rsubread::align() function, which is called inside align_target(). Elements should include type, nthreads, maxMismatches, nsubreads, phredOffset, unique, and nBestLocations. Descriptions of these parameters are available under ?Rsubread::align. Defaults to the global align_details object.

Value

This function writes a merged and sorted .bam file after aligning to all reference libraries given, along with a summary report file, to the user's working directory. The function also outputs the new .bam filename.

Examples

#### Align example reads to an example reference library using Rsubread
# \donttest{
## Create temporary directory
target_ref_temp <- tempfile()
dir.create(target_ref_temp)

## Download genome
tax <- "Mycoplasma genitalium G37"
all_ref <- MetaScope::download_refseq(tax,
                                      reference = FALSE,
                                      representative = FALSE,
                                      compress = TRUE,
                                      out_dir = target_ref_temp
)
#> Finding Mycoplasma genitalium G37
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> Mycoplasma genitalium G37 is a strain under the Bacteria Superkingdom
#> Loading the refseq table for bacteria
#> Creating table of relevant taxa
#> Downloading 2 Mycoplasma genitalium G37 genome(s) from NCBI
#> DONE! Downloaded 2 genomes to /scratch/141165.1.ood/RtmpwygBTw/file7c3324afa3a5//scratch/141165.1.ood/RtmpwygBTw/file7c3324afa3a5/Mycoplasma_genitalium_G37.fasta.gz

## Create subread index
ind_out <- mk_subread_index(all_ref)
#> Library size is 0 GBs. Building the Rsubread index
#> 
#>         ==========     _____ _    _ ____  _____  ______          _____  
#>         =====         / ____| |  | |  _ \|  __ \|  ____|   /\   |  __ \ 
#>           =====      | (___ | |  | | |_) | |__) | |__     /  \  | |  | |
#>             ====      \___ \| |  | |  _ <|  _  /|  __|   / /\ \ | |  | |
#>               ====    ____) | |__| | |_) | | \ \| |____ / ____ \| |__| |
#>         ==========   |_____/ \____/|____/|_|  \_\______/_/    \_\_____/
#>        Rsubread 2.10.5
#> 
#> //================================= setting ==================================\\
#> ||                                                                            ||
#> ||                Index name : Mycoplasma_genitalium_G37                      ||
#> ||               Index space : base space                                     ||
#> ||               Index split : no-split                                       ||
#> ||          Repeat threshold : 100 repeats                                    ||
#> ||              Gapped index : no                                             ||
#> ||                                                                            ||
#> ||       Free / total memory : 200.5GB / 251.7GB                              ||
#> ||                                                                            ||
#> ||               Input files : 1 file in total                                ||
#> ||                             o Mycoplasma_genitalium_G37.fasta.gz           ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> //================================= Running ==================================\\
#> ||                                                                            ||
#> || Check the integrity of provided reference sequences ...                    ||
#> || No format issues were found                                                ||
#> || Scan uninformative subreads in reference sequences ...                     ||
#> || Estimate the index size...                                                 ||
#> ||    8%,   0 mins elapsed, rate=491.0k bps/s                                 ||
#> ||   16%,   0 mins elapsed, rate=914.5k bps/s                                 ||
#> ||   24%,   0 mins elapsed, rate=1284.4k bps/s                                ||
#> ||   33%,   0 mins elapsed, rate=1594.4k bps/s                                ||
#> ||   41%,   0 mins elapsed, rate=1881.4k bps/s                                ||
#> ||   58%,   0 mins elapsed, rate=2450.5k bps/s                                ||
#> ||   66%,   0 mins elapsed, rate=2703.6k bps/s                                ||
#> ||   74%,   0 mins elapsed, rate=2944.2k bps/s                                ||
#> ||   83%,   0 mins elapsed, rate=3156.9k bps/s                                ||
#> ||   91%,   0 mins elapsed, rate=3337.6k bps/s                                ||
#> || 3.0 GB of memory is needed for index building.                             ||
#> || Build the index...                                                         ||
#> ||    8%,   0 mins elapsed, rate=49.5k bps/s                                  ||
#> ||   16%,   0 mins elapsed, rate=97.5k bps/s                                  ||
#> ||   24%,   0 mins elapsed, rate=144.4k bps/s                                 ||
#> ||   33%,   0 mins elapsed, rate=190.0k bps/s                                 ||
#> ||   41%,   0 mins elapsed, rate=234.3k bps/s                                 ||
#> ||   58%,   0 mins elapsed, rate=321.4k bps/s                                 ||
#> ||   66%,   0 mins elapsed, rate=363.0k bps/s                                 ||
#> ||   74%,   0 mins elapsed, rate=403.8k bps/s                                 ||
#> ||   83%,   0 mins elapsed, rate=443.5k bps/s                                 ||
#> ||   91%,   0 mins elapsed, rate=482.3k bps/s                                 ||
#> || Save current index block...                                                ||
#> ||  [ 0.0% finished ]                                                         ||
#> ||  [ 10.0% finished ]                                                        ||
#> ||  [ 20.0% finished ]                                                        ||
#> ||  [ 30.0% finished ]                                                        ||
#> ||  [ 40.0% finished ]                                                        ||
#> ||  [ 50.0% finished ]                                                        ||
#> ||  [ 60.0% finished ]                                                        ||
#> ||  [ 70.0% finished ]                                                        ||
#> ||  [ 80.0% finished ]                                                        ||
#> ||  [ 90.0% finished ]                                                        ||
#> ||  [ 100.0% finished ]                                                       ||
#> ||                                                                            ||
#> ||                      Total running time: 0.2 minutes.                      ||
#> ||Index /scratch/141165.1.ood/RtmpwygBTw/file7c3324afa3a5/Mycoplasma_gen ... ||
#> ||                                                                            ||
#> \\============================================================================//
#> 

## Get path to example reads
readPath <- system.file("extdata", "reads.fastq",
                        package = "MetaScope")
## Copy the example reads to the temp directory
refPath <- file.path(target_ref_temp, "reads.fastq")
file.copy(from = readPath, to = refPath)
#> [1] TRUE

## Modify alignment parameters object
data("align_details")
align_details[["type"]] <- "rna"
align_details[["phredOffset"]] <- 50
# Just to get it to align - toy example!
align_details[["maxMismatches"]] <- 100

## Run alignment
target_map <- align_target(refPath,
                           libs = stringr::str_replace_all(tax, " ", "_"),
                           lib_dir = target_ref_temp,
                           subread_options = align_details)
#> Attempting to perform subread alignment on Mycoplasma_genitalium_G37 index
#> 
#>         ==========     _____ _    _ ____  _____  ______          _____  
#>         =====         / ____| |  | |  _ \|  __ \|  ____|   /\   |  __ \ 
#>           =====      | (___ | |  | | |_) | |__) | |__     /  \  | |  | |
#>             ====      \___ \| |  | |  _ <|  _  /|  __|   / /\ \ | |  | |
#>               ====    ____) | |__| | |_) | | \ \| |____ / ____ \| |__| |
#>         ==========   |_____/ \____/|____/|_|  \_\______/_/    \_\_____/
#>        Rsubread 2.10.5
#> 
#> //================================= setting ==================================\\
#> ||                                                                            ||
#> || Function      : Read alignment (RNA-Seq)                                   ||
#> || Input file    : reads.fastq                                                ||
#> || Output file   : reads.Mycoplasma_genitalium_G37.bam (BAM)                  ||
#> || Index name    : Mycoplasma_genitalium_G37                                  ||
#> ||                                                                            ||
#> ||                    ------------------------------------                    ||
#> ||                                                                            ||
#> ||                               Threads : 1                                  ||
#> ||                          Phred offset : 64                                 ||
#> ||                             Min votes : 3 / 10                             ||
#> ||                        Max mismatches : 100                                ||
#> ||                      Max indel length : 5                                  ||
#> ||            Report multi-mapping reads : yes                                ||
#> || Max alignments per multi-mapping read : 16                                 ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> //================ Running (28-Oct-2022 16:42:15, pid=31795) =================\\
#> ||                                                                            ||
#> || Check the input reads.                                                     ||
#> || The input file contains base space reads.                                  ||
#> || Initialise the memory objects.                                             ||
#> || Estimate the mean read length.                                             ||
#> || The range of Phred scores observed in the data is [-14,-14]                ||
#> || Create the output BAM file.                                                ||
#> || Check the index.                                                           ||
#> || Init the voting space.                                                     ||
#> || Global environment is initialised.                                         ||
#> || Load the 1-th index block...                                               ||
#> || The index block has been loaded.                                           ||
#> || Start read mapping in chunk.                                               ||
#> ||                                                                            ||
#> ||                           Completed successfully.                          ||
#> ||                                                                            ||
#> \\====================================    ====================================//
#> 
#> //================================   Summary =================================\\
#> ||                                                                            ||
#> ||                 Total reads : 1,500                                        ||
#> ||                      Mapped : 3 (0.2%)                                     ||
#> ||             Uniquely mapped : 0                                            ||
#> ||               Multi-mapping : 3                                            ||
#> ||                                                                            ||
#> ||                    Unmapped : 1,497                                        ||
#> ||                                                                            ||
#> ||                      Indels : 0                                            ||
#> ||                                                                            ||
#> ||                Running time : 1.0 minutes                                  ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> Filtering unmapped reads
#> Library alignment complete
#> DONE! Alignments written to /scratch/141165.1.ood/RtmpwygBTw/file7c3324afa3a5/reads.bam

## Remove temporary folder
unlink(target_ref_temp, recursive = TRUE)
# }