Download RefSeq genome libraries

This function will automatically download RefSeq genome libraries in a fasta format from the specified taxon. The function will first download the summary report at: ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/**kingdom**/assembly_summary.txt, and then use this file to download the genome(s) and combine them in a single compressed or uncompressed .fasta file.

Usage

download_refseq(
  taxon,
  reference = TRUE,
  representative = FALSE,
  compress = TRUE,
  patho_out = FALSE,
  out_dir = NULL
)

Arguments

taxon

Name of single taxon to download. The taxon name should be a recognized NCBI scientific or common name, with no grammatical or capitalization inconsistencies. All available taxonomies are visible by accessing the taxonomy_table object included in the package.

reference

Download only RefSeq reference genomes? Defaults to TRUE. Automatically set to TRUE if representative = TRUE.

representative

Download only RefSeq representative genomes? Defaults to FALSE. If TRUE, reference is automatically set at TRUE.

compress

Compress the output .fasta file? Defaults to TRUE.

patho_out

Create duplicate outpute files compatible with PathoScope? Defaults to FALSE.

out_dir

Character string giving the name of the directory to which libraries should be output.

Value

Returns a .fasta or .fasta.gz file of the desired RefSeq genomes. This file is named after the kingdom selected and saved to the current directory (e.g. 'bacteria.fasta.gz'). This function also has the option to return a .fasta file formatted for PathoScope as well (e.g. bacteria.pathoscope.fasta.gz') if path_out = TRUE.

Details

When selecting the taxon to be downloaded, if you receive an error saying Your input is not a valid taxon, please take a look at the taxonomy_table object, which can be accessed with the command data("taxonomy_table"). Only taxa with exact spelling as they appear at any level of the table will be acknowledged.

Examples

#### Download RefSeq genomes
## Create a temporary directory to store libraries
ref_temp <- tempfile()
dir.create(ref_temp)

## Download all RefSeq reference Shotokuvirae kingdom genomes
download_refseq('Shotokuvirae', reference = TRUE, representative = FALSE,
                out_dir = ref_temp, compress = TRUE, patho_out = FALSE)
#> Finding Shotokuvirae
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> Shotokuvirae is a kingdom under the Viruses Superkingdom
#> Loading the refseq table for viral
#> Creating table of relevant taxa
#> Downloading 1 Shotokuvirae genome(s) from NCBI
#> DONE! Downloaded 1 genomes to /scratch/141165.1.ood/RtmpwygBTw/file7c3366b8d5c0//scratch/141165.1.ood/RtmpwygBTw/file7c3366b8d5c0/Shotokuvirae.fasta.gz
#> [1] "/scratch/141165.1.ood/RtmpwygBTw/file7c3366b8d5c0/Shotokuvirae.fasta.gz"

unlink(ref_temp)