Align reads against one or more filter libraries and subsequently remove mapped reads

After a sample is aligned to a target library with align_target(), we may use filter_host() to remove unwelcome host contamination using filter reference libraries. This function takes as input the name of the .bam file produced via align_target(), and produces a sorted .bam file with any reads that match the filter libraries removed. This resulting .bam file may be used upstream for further analysis.

Usage

filter_host(
  reads_bam,
  libs,
  lib_dir = NULL,
  output = paste(tools::file_path_sans_ext(reads_bam), "filtered", "bam", sep = "."),
  settings = align_details
)

Arguments

reads_bam

The name of a merged, sorted .bam file that has previously been aligned to a reference library. Likely, the output from running an instance of align_target().

libs

A vector of character strings giving the basenames of the Subread index files for alignment. These should be located in the same directory and the current directory if lib_dir=NULL.

lib_dir

path to the library index files (all libraries should be here)

output

The desired name of the output .bam file. Default is the basename of reads_bam + .filtered.bam.

settings

A named list specifying alignment parameters for the Rsubread::align() function, which is called inside align_target(). Elements should include type, nthreads, maxMismatches, nsubreads, phredOffset, unique, and nBestLocations. Descriptions of these parameters are available under ?Rsubread::align. Defaults to the global align_details object.

Value

The name of a filtered, sorted .bam file written to the user's current working directory.

Examples

#### Filter reads from bam file that align to any of the filter libraries

## Assuming a bam file has been created previously with align_target()

## Create object with path to the example filter library
refPath <- system.file("extdata","filter.fasta", package = "MetaScope")

## Copy the example filter library to the current directory
file.copy(from = refPath, to = file.path(".", "filter.fasta"))
#> [1] FALSE

## Make subread index of filter library
mk_subread_index('filter.fasta')
#> [1] "Library size is 0 GBs. Building the Rsubread index"
#> 
#>         ==========     _____ _    _ ____  _____  ______          _____  
#>         =====         / ____| |  | |  _ \|  __ \|  ____|   /\   |  __ \ 
#>           =====      | (___ | |  | | |_) | |__) | |__     /  \  | |  | |
#>             ====      \___ \| |  | |  _ <|  _  /|  __|   / /\ \ | |  | |
#>               ====    ____) | |__| | |_) | | \ \| |____ / ____ \| |__| |
#>         ==========   |_____/ \____/|____/|_|  \_\______/_/    \_\_____/
#>        Rsubread 2.8.1
#> 
#> //================================= setting ==================================\\
#> ||                                                                            ||
#> ||                Index name : filter                                         ||
#> ||               Index space : base space                                     ||
#> ||               Index split : no-split                                       ||
#> ||          Repeat threshold : 100 repeats                                    ||
#> ||              Gapped index : no                                             ||
#> ||                                                                            ||
#> ||       Free / total memory : 317.5GB / 376.4GB                              ||
#> ||                                                                            ||
#> ||               Input files : 1 file in total                                ||
#> ||                             o filter.fasta                                 ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> //================================= Running ==================================\\
#> ||                                                                            ||
#> || Check the integrity of provided reference sequences ...                    ||
#> || No format issues were found                                                ||
#> || Scan uninformative subreads in reference sequences ...                     ||
#> || Estimate the index size...                                                 ||
#> ||    8%,   0 mins elapsed, rate=1174.7k bps/s                                ||
#> ||   16%,   0 mins elapsed, rate=1946.2k bps/s                                ||
#> ||   24%,   0 mins elapsed, rate=2544.9k bps/s                                ||
#> ||   33%,   0 mins elapsed, rate=3007.1k bps/s                                ||
#> ||   41%,   0 mins elapsed, rate=3376.2k bps/s                                ||
#> ||   49%,   0 mins elapsed, rate=3678.6k bps/s                                ||
#> ||   58%,   0 mins elapsed, rate=3930.2k bps/s                                ||
#> ||   66%,   0 mins elapsed, rate=4143.1k bps/s                                ||
#> ||   74%,   0 mins elapsed, rate=4322.9k bps/s                                ||
#> ||   83%,   0 mins elapsed, rate=4433.2k bps/s                                ||
#> ||   91%,   0 mins elapsed, rate=4577.4k bps/s                                ||
#> || 3.0 GB of memory is needed for index building.                             ||
#> || Build the index...                                                         ||
#> ||    8%,   0 mins elapsed, rate=131.0k bps/s                                 ||
#> ||   16%,   0 mins elapsed, rate=251.3k bps/s                                 ||
#> ||   24%,   0 mins elapsed, rate=362.0k bps/s                                 ||
#> ||   33%,   0 mins elapsed, rate=463.9k bps/s                                 ||
#> ||   41%,   0 mins elapsed, rate=558.4k bps/s                                 ||
#> ||   49%,   0 mins elapsed, rate=646.3k bps/s                                 ||
#> ||   58%,   0 mins elapsed, rate=728.2k bps/s                                 ||
#> ||   66%,   0 mins elapsed, rate=804.9k bps/s                                 ||
#> ||   74%,   0 mins elapsed, rate=876.4k bps/s                                 ||
#> ||   83%,   0 mins elapsed, rate=943.5k bps/s                                 ||
#> ||   91%,   0 mins elapsed, rate=1005.2k bps/s                                ||
#> || Save current index block...                                                ||
#> ||  [ 0.0% finished ]                                                         ||
#> ||  [ 10.0% finished ]                                                        ||
#> ||  [ 20.0% finished ]                                                        ||
#> ||  [ 30.0% finished ]                                                        ||
#> ||  [ 40.0% finished ]                                                        ||
#> ||  [ 50.0% finished ]                                                        ||
#> ||  [ 60.0% finished ]                                                        ||
#> ||  [ 70.0% finished ]                                                        ||
#> ||  [ 80.0% finished ]                                                        ||
#> ||  [ 90.0% finished ]                                                        ||
#> ||  [ 100.0% finished ]                                                       ||
#> ||                                                                            ||
#> ||                      Total running time: 0.2 minutes.                      ||
#> ||                    Index filter was successfully built.                    ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> [1] "filter_1"

## Create object with path to the previously aligned bam file
readPath <- system.file("extdata", "subread_target.bam", package = "MetaScope")

## Filter bam file 
filter_host(readPath, libs = "filter")
#> 
#>         ==========     _____ _    _ ____  _____  ______          _____  
#>         =====         / ____| |  | |  _ \|  __ \|  ____|   /\   |  __ \ 
#>           =====      | (___ | |  | | |_) | |__) | |__     /  \  | |  | |
#>             ====      \___ \| |  | |  _ <|  _  /|  __|   / /\ \ | |  | |
#>               ====    ____) | |__| | |_) | | \ \| |____ / ____ \| |__| |
#>         ==========   |_____/ \____/|____/|_|  \_\______/_/    \_\_____/
#>        Rsubread 2.8.1
#> 
#> //================================= setting ==================================\\
#> ||                                                                            ||
#> || Function      : Read alignment (DNA-Seq)                                   ||
#> || Input file    : subread_target.bam (BAM)                                   ||
#> || Output file   : subread_target.filter.bam (BAM)                            ||
#> || Index name    : filter                                                     ||
#> ||                                                                            ||
#> ||                    ------------------------------------                    ||
#> ||                                                                            ||
#> ||                               Threads : 8                                  ||
#> ||                          Phred offset : 33                                 ||
#> ||                             Min votes : 3 / 10                             ||
#> ||                        Max mismatches : 3                                  ||
#> ||                      Max indel length : 5                                  ||
#> ||            Report multi-mapping reads : yes                                ||
#> || Max alignments per multi-mapping read : 16                                 ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> //================ Running (11-Mar-2022 14:00:06, pid=120161) ================\\
#> ||                                                                            ||
#> || Check the input reads.                                                     ||
#> || The input BAM file contains single-end reads.                              ||
#> || Convert the input BAM file...                                              ||
#> || The input file contains base space reads.                                  ||
#> || Initialise the memory objects.                                             ||
#> || Estimate the mean read length.                                             ||
#> || WARNING  - The specified Phred score offset (33) seems incorrect.          ||
#> ||            ASCII values of the quality scores of read bases included in    ||
#> ||            the first 2490 reads were found to be within the range of       ||
#> ||            [50,50].                                                        ||
#> ||                                                                            ||
#> || Create the output BAM file.                                                ||
#> || Check the index.                                                           ||
#> || Init the voting space.                                                     ||
#> || Global environment is initialised.                                         ||
#> || Load the 1-th index block...                                               ||
#> || The index block has been loaded.                                           ||
#> || Start read mapping in chunk.                                               ||
#> ||                                                                            ||
#> ||                           Completed successfully.                          ||
#> ||                                                                            ||
#> \\====================================    ====================================//
#> 
#> //================================   Summary =================================\\
#> ||                                                                            ||
#> ||                 Total reads : 1,491                                        ||
#> ||                      Mapped : 94 (6.3%)                                    ||
#> ||             Uniquely mapped : 73                                           ||
#> ||               Multi-mapping : 21                                           ||
#> ||                                                                            ||
#> ||                    Unmapped : 1,397                                        ||
#> ||                                                                            ||
#> ||                      Indels : 7                                            ||
#> ||                                                                            ||
#> ||                     WARNING : Phred offset (33) incorrect?                 ||
#> ||                                                                            ||
#> ||                Running time : 0.7 minutes                                  ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> DONE! Alignments written to /scratch/3276553.1.linga/Rtmpp8sDKo/temp_libpath1c89f2e042019/MetaScope/extdata/subread_target.filtered.bam
#> [1] "/scratch/3276553.1.linga/Rtmpp8sDKo/temp_libpath1c89f2e042019/MetaScope/extdata/subread_target.filtered.bam"