Align reads against one or more filter libraries and subsequently remove mapped reads

After a sample is aligned to a target library with align_target_bowtie(), we may use filter_host_bowtie() to remove unwelcome host contamination using filter reference libraries. This function takes as input the name of the .bam file produced via align_target_bowtie(), and produces a sorted .bam file with any reads that match the filter libraries removed. This resulting .bam file may be used downstream for further analysis.

Usage

filter_host_bowtie(
  reads_bam,
  lib_dir,
  libs,
  output = paste(tools::file_path_sans_ext(reads_bam), "filtered", "bam", sep = "."),
  bowtie2_options = NULL,
  threads = 8,
  overwrite = FALSE,
  maxMemory = 512
)

Arguments

reads_bam

The name of a merged, sorted .bam file that has previously been aligned to a reference library. Likely, the output from running an instance of align_target_bowtie().

lib_dir

Path to the directory that contains the filter Bowtie2 index files.

libs

The basename of the filter libraries (without .bt2 or .bt2l extension)

output

The desired name of the output .bam file. Default is the basename of unfiltered_bam + .filtered.bam.

bowtie2_options

Optional: Additional parameters that can be passed to the filter_host_bowtie() function. To see all the available parameters use Rbowtie2::bowtie2_usage(). Default parameters are the parameters are the default parameters that PathoScope 2.0 uses. NOTE: Users should pass all their parameters as one string and if optional parameters are given then the user is responsible for entering all the parameters to be used by Bowtie2. NOTE: The only parameters that should NOT be specified here is the threads.

threads

The amount of threads available for the function. Default is 8 threads.

overwrite

Whether existing files should be overwritten. Default is FALSE.

maxMemory

The number of MB of RAM that the sortBAM function can use. The smaller the number, the more temporary files will be produced, which will hopefully save memory usage. The default is 512 MB.

Value

The name of a filtered, sorted .bam file written to the user's current working directory.

Examples

#### Filter reads from bam file that align to any of the filter libraries

## Assuming a bam file has already been created with align_target_bowtie()

## Create a temporary directory to store the filter library
ref_temp <- tempfile()
dir.create(ref_temp)

## Create a temporary directory to store the filter library index files
lib_temp <- tempfile()
dir.create(lib_temp)

## Create a temporary directory to store the filtered bam file
align_temp <- tempfile()
dir.create(align_temp)

## Create object with path to previously created bam file 
bamPath <- system.file("extdata", "bowtie_target.bam", package = "MetaScope")

## Create object with path to the filter library
refPath <- system.file("extdata","filter.fasta", package = "MetaScope")

## Move the filter library to the temporary reference directory 
file.copy(from = refPath, to = file.path(ref_temp, "filter.fasta"))
#> [1] TRUE

## Create the bowtie index files in the temporary index directory
mk_bowtie_index(ref_dir = ref_temp, lib_dir = lib_temp, lib_name = "filter",
overwrite=FALSE)
#> arguments 'show.output.on.console', 'minimized' and 'invisible' are for Windows only
#> Successfully built the Bowtie2 indexes
#> [1] "/scratch/3276553.1.linga/RtmpEXvmSK/file1ee2c61b5fd"

## Filter reads from the bam file that align to the filter library
filter_host_bowtie(reads_bam = bamPath, lib_dir = lib_temp, libs = "filter")
#> Creating Intermediate Fastq file
#> Finished Creating Intermediate Fastq File
#> [1] "Samtools not found on system. Using Rsamtools to create bam file"
#> arguments 'show.output.on.console', 'minimized' and 'invisible' are for Windows only
#> DONE! Alignments written to /scratch/3276553.1.linga/Rtmph47wXA/temp_libpath1d6ac7e37a51d/MetaScope/extdata/bowtie_target.filtered.bam
#> [1] "/scratch/3276553.1.linga/Rtmph47wXA/temp_libpath1d6ac7e37a51d/MetaScope/extdata/bowtie_target.filtered.bam"