MetaScope ID

This function will read in a .bam file, annotate the taxonomy and genome names, reduce the mapping ambiguity using a mixture model, and output a .csv file with the results. Currently, it assumes that the genome library/.bam files use NCBI accession names for reference names (rnames in .bam file).

Usage

metascope_id(
  bam_file,
  aligner = "subread",
  NCBI_key = NULL,
  out_file = paste(tools::file_path_sans_ext(bam_file), ".metascope_id.csv", sep = ""),
  EMconv = 1/10000,
  EMmaxIts = 25,
  num_species_plot = NULL
)

Arguments

bam_file

The .bam file that needs to be summarized, annotated, and needs removal of ambiguity.

aligner

The aligner which was used to create the bam file. Default is "subread" but can also be set to "bowtie" or "other"

NCBI_key

(character) NCBI Entrez API key. optional. See taxize::use_entrez(). Due to the high number of requests made to NCBI, this function will be less prone to errors if you obtain an NCBI key. You may enter the string as an input or set it as ENTREZ_KEY in .Renviron.

out_file

The name of the .csv output file. Defaults to the bam_file basename plus ".metascope_id.csv".

EMconv

The convergence parameter of the EM algorithm. Default set at 1/10000.

EMmaxIts

The maximum number of EM iterations, regardless of whether the EMconv is below the threshhold. Default set at 50. If set at 0, the algorithm skips the EM step and summarizes the .bam file 'as is'

num_species_plot

The number of genome coverage plots to be saved. Default is NULL, which saves coverage plots for the ten most highly abundant species.

Value

This function returns a .csv file with annotated read counts to genomes with mapped reads. The function itself returns the output .csv file name.

Examples

#### Align reads to reference library and then apply metascope_id()

## Assuming filtered bam files already exist

## Subread aligned bam file

## Create object with path to filtered subread bam file 
bamPath <- system.file("extdata","subread_target.filtered.bam",
package = "MetaScope")

## Run metascope id with the aligner option set to subread
metascope_id(bam_file = bamPath, aligner = "subread")
#> Reading .bam file: /scratch/3276553.1.linga/Rtmph47wXA/temp_libpath1d6ac7e37a51d/MetaScope/extdata/subread_target.filtered.bam
#> 	Found 856 reads aligned to 77 NCBI accessions
#> Obtaining taxonomy and genome names
#> 	Found 6 unique NCBI taxonomy IDs
#> Setting up the EM algorithm
#> Starting EM iterations
#> [1]   1.0000 812.7488
#> [1] 2.0000000 0.8139525
#> [1] 3.0000000 0.4173796
#> [1] 4.0000000 0.5594038
#> [1] 5.0000000 0.6864919
#> [1] 6.0000000 0.7528093
#> [1] 7.0000000 0.7763587
#> [1] 8.0000000 0.7836479
#> [1] 9.0000000 0.7861907
#> [1] 10.0000000  0.7873258
#> [1] 11.0000000  0.7879545
#> [1] 12.0000000  0.7883482
#> [1] 13.000000  0.788609
#> [1] 14.000000  0.788786
#> [1] 15.0000000  0.7889075
#> [1] 16.0000000  0.7889914
#> [1] 17.0000000  0.7890495
#> [1] 18.00000  0.78909
#> [1] 19.0000000  0.7891182
#> [1] 20.0000000  0.7891379
#> [1] 21.0000000  0.7891516
#> [1] 22.0000000  0.7891613
#> [1] 23.0000000  0.7891681
#> [1] 24.0000000  0.7891729
#> [1] 25.0000000  0.7891762
#> 	DONE! Converged in 25 interations.
#> Found reads for 5 genomes
#> Creating coverage plots
#> Saving 6.67 x 6.67 in image
#> Saving 6.67 x 6.67 in image
#> Saving 6.67 x 6.67 in image
#> Saving 6.67 x 6.67 in image
#> Saving 6.67 x 6.67 in image
#> Results written to /scratch/3276553.1.linga/Rtmph47wXA/temp_libpath1d6ac7e37a51d/MetaScope/extdata/subread_target.filtered.metascope_id.csv
#>   TaxonomyID
#> 1     273036
#> 2     418127
#> 3     523796
#> 4     426430
#> 5     158878
#>                                                                                          Genome
#> 1                                                Staphylococcus aureus RF122, complete sequence
#> 2                        Staphylococcus aureus subsp. aureus Mu3, whole genome shotgun sequence
#> 3 Staphylococcus aureus subsp. aureus ST398 strain 55488 NODE_10, whole genome shotgun sequence
#> 4                            Staphylococcus aureus subsp. aureus str. Newman, complete sequence
#> 5                                   Staphylococcus aureus subsp. aureus Mu50, complete sequence
#>   read_count          Proportion EMreads        EMProportion
#> 1        844   0.985981308411215     844   0.985968411207406
#> 2          6 0.00700934579439252     3.8 0.00440144211459077
#> 3          2 0.00233644859813084     2.3 0.00271337559673324
#> 4          2 0.00233644859813084     2.2 0.00257358934601975
#> 5          2 0.00233644859813084     3.7 0.00434318173524827

## Bowtie aligned bam file 

## Create object with path to filtered subread bam file 
bamPath <- system.file("extdata","bowtie_target.filtered.bam",
package = "MetaScope")

## Run metascope id with the aligner option set to bowtie
metascope_id(bam_file = bamPath, aligner = "bowtie")
#> Reading .bam file: /scratch/3276553.1.linga/Rtmph47wXA/temp_libpath1d6ac7e37a51d/MetaScope/extdata/bowtie_target.filtered.bam
#> 	Found 955 reads aligned to 120 NCBI accessions
#> Obtaining taxonomy and genome names
#> 	Found 6 unique NCBI taxonomy IDs
#> Setting up the EM algorithm
#> Starting EM iterations
#> [1]   1.0000 884.5066
#> [1] 2.0000000 0.9827828
#> [1] 3.0000000 0.7919072
#> [1] 4.0000000 0.7871586
#> [1] 5.0000000 0.8068483
#> [1] 6.0000000 0.8193158
#> [1] 7.0000000 0.8275384
#> [1] 8.0000000 0.8332668
#> [1] 9.0000000 0.8372379
#> [1] 10.000000  0.839936
#> [1] 11.0000000  0.8417388
#> [1] 12.0000000  0.8429307
#> [1] 13.0000000  0.8437146
#> [1] 14.0000000  0.8442325
#> [1] 15.0000000  0.8445838
#> [1] 16.0000000  0.8448411
#> [1] 17.0000000  0.8450599
#> [1] 18.0000000  0.8452807
#> [1] 19.0000000  0.8455173
#> [1] 20.0000000  0.8457242
#> [1] 21.0000000  0.8457509
#> [1] 22.0000000  0.8453414
#> [1] 23.0000000  0.8442443
#> [1] 24.000000  0.842381
#> [1] 25.0000000  0.8398935
#> 	DONE! Converged in 25 interations.
#> Found reads for 3 genomes
#> Creating coverage plots
#> Saving 6.67 x 6.67 in image
#> Saving 6.67 x 6.67 in image
#> Saving 6.67 x 6.67 in image
#> Results written to /scratch/3276553.1.linga/Rtmph47wXA/temp_libpath1d6ac7e37a51d/MetaScope/extdata/bowtie_target.filtered.metascope_id.csv
#>   TaxonomyID                                                             Genome
#> 1     273036                     Staphylococcus aureus RF122, complete sequence
#> 3     158878        Staphylococcus aureus subsp. aureus Mu50, complete sequence
#> 6     426430 Staphylococcus aureus subsp. aureus str. Newman, complete sequence
#>   read_count          Proportion EMreads         EMProportion
#> 1        953   0.997905759162304   953.1    0.998026139867055
#> 3          1 0.00104712041884817     1.3  0.00131623459197387
#> 6          1 0.00104712041884817     0.6 0.000625302300008886

## Different or unknown aligned bam file

## Create object with path to unknown origin bam file 
bamPath <- system.file("extdata","subread_target.filtered.bam",
package = "MetaScope")

## Run metascope id with the aligner option set to other 
metascope_id(bam_file = bamPath, aligner = "other")
#> Reading .bam file: /scratch/3276553.1.linga/Rtmph47wXA/temp_libpath1d6ac7e37a51d/MetaScope/extdata/subread_target.filtered.bam
#> 	Found 856 reads aligned to 77 NCBI accessions
#> Obtaining taxonomy and genome names
#> 	Found 6 unique NCBI taxonomy IDs
#> Setting up the EM algorithm
#> Starting EM iterations
#> [1]   1.0000 759.5111
#> [1] 2.0000000 0.9132601
#> [1] 3.0000000 0.4755713
#> [1] 4.000000 0.568628
#> [1] 5.0000000 0.6922415
#> [1] 6.0000000 0.7547038
#> [1] 7.0000000 0.7766409
#> [1] 8.0000000 0.7835835
#> [1] 9.0000000 0.7861197
#> [1] 10.000000  0.787304
#> [1] 11.0000000  0.7879779
#> [1] 12.0000000  0.7884051
#> [1] 13.0000000  0.7886894
#> [1] 14.000000  0.788883
#> [1] 15.0000000  0.7890162
#> [1] 16.0000000  0.7891085
#> [1] 17.0000000  0.7891726
#> [1] 18.0000000  0.7892174
#> [1] 19.0000000  0.7892487
#> [1] 20.0000000  0.7892707
#> [1] 21.0000000  0.7892861
#> [1] 22.000000  0.789297
#> [1] 23.0000000  0.7893047
#> [1] 24.0000000  0.7893102
#> [1] 25.0000000  0.7893141
#> 	DONE! Converged in 25 interations.
#> Found reads for 5 genomes
#> Creating coverage plots
#> Saving 6.67 x 6.67 in image
#> Saving 6.67 x 6.67 in image
#> Saving 6.67 x 6.67 in image
#> Saving 6.67 x 6.67 in image
#> Saving 6.67 x 6.67 in image
#> Results written to /scratch/3276553.1.linga/Rtmph47wXA/temp_libpath1d6ac7e37a51d/MetaScope/extdata/subread_target.filtered.metascope_id.csv
#>   TaxonomyID
#> 1     273036
#> 2     418127
#> 3     523796
#> 4     426430
#> 5     158878
#>                                                                                          Genome
#> 1                                                Staphylococcus aureus RF122, complete sequence
#> 2                        Staphylococcus aureus subsp. aureus Mu3, whole genome shotgun sequence
#> 3 Staphylococcus aureus subsp. aureus ST398 strain 55488 NODE_10, whole genome shotgun sequence
#> 4                            Staphylococcus aureus subsp. aureus str. Newman, complete sequence
#> 5                                   Staphylococcus aureus subsp. aureus Mu50, complete sequence
#>   read_count          Proportion EMreads        EMProportion
#> 1        844   0.985981308411215     844   0.985960405472934
#> 2          6 0.00700934579439252     3.8 0.00440897491933567
#> 3          2 0.00233644859813084     2.3 0.00271433413968925
#> 4          2 0.00233644859813084     2.2 0.00257414474820706
#> 5          2 0.00233644859813084     3.7 0.00434214071983583